dicetower/dice_detection_server_requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
# Uncomment if using TensorFlow model:
# tensorflow>=2.13.0
# keras>=2.13.0

